Delete Ryedale import script
Recent interface changes make it clear that the shapes provided
by Ryedale are not actually polling districts. Most of them
contain more than one polling station and I'm not confident that
the ones that contain one station are actually the polling station
for _everyone_ in that area - so remove this on the basis it
provides no value.