Retry opening KML after initial error
As per the comment in the code, this is a very odd bug, with a poor fix!

The first time I try to open a KML file I get an "unknown error" exceptoion raises, with a different error number each time. Seems like trying to open a KML file twice works on GDAL 2.2.0 / OS X.