Stop relying on update_or_create() to implicitly de-dupe
This slightly improves performance as well, but the main focus
of this is data integrity. If there are duplicates that violate
the unique constraint (council_id, internal_council_id) in the
PollingStation/PollingDistrict models, this should throw an exception
rather than silently overwriting with the last record encountered.

In the case where there are duplicates on
(council_id, internal_council_id) we should explicitly handle
this in an appropriate manner in our import scripts.

Although in some cases, we will need to essentially replicate this
behaviour, making this change highlights some situations where use
of update_or_create() actually introduced bugs/inaccuracies.