Add generate_csvs script
This uses the ElectoralRoll source data to generate CSV
templates. These templates are then distributed to wombles
who can fill in the polling station addreses and send the
completed CSVs back to us for import

For the moment we will abstract the issue of populating
the ElectoralRoll table with data in he first place.