Allow data to be fetched from S3 using an arbitary prefix
Usually data is sourced from councils and will be stored
on S3 in a dir named with local authority GSS code.

Data from the Scotland SpatialHub will be provided in a single
dataset for the whole country. As such, we need to be able
to store data on S3 which is not keyed by local auth code.