Add ExamplePostcodeHelper
This adds a helper and context processor for defining an example
postcode that will be displayed on the home page.

It's a class so we have more ability in future to do more complex
things, for example by settings a different postcode for white
label or embed sites, or dynamically setting a postcode based on
imported data.

Refs #575