Halt on exception, print useful stacktrace
This prints a stacktrace that actually tells us where in the offending
import script the exception was raised (even when running in parallel).
I think in general we don't want to skip over exceptions as depending
where the exception is thrown, passing on to the next import script may
leave our DB in an undesired state
(for example if the exception was an `IntegrityError`).