Ensure postcode is correctly formatted before geocode from AddressBase
Whether we geocode from AddressBase or fall back to mapit was
partly dependent on whether or not the postcode contained a space.

In particular, if we'd already stripped the whitespace from a
postcode and then called `geocode_point_only(postcode)`
(like we do in the /address API endpoint code) we would always
fail to match anything in AddressBase and fall back to mapit.

This fixes that issue.