Improve import script performance
Previously EdgeCaseFixer.make_addresses_for_postcode() was running
a minimum of 2 DB queries per address - sometimes more.
Now we perform one DB query per postcode which returns all the
information we need for all the addresses in that postcode.
This massively reduces the amount of time we spend performing
database I/O in this phase of the import.