Handle polling districts that don't exactly match council boundary
If 2 neighbouring councils provide slightly overlapping polygons,
this was causing a problem because we were sometimes associating
the same address with more than one council in
make_addresses_for_postcode().

This commit fixes this by determining the council area based on
looking up the address point in the councils table instead of
inferring it from the polling district.