Allow validation checks to be skipped
Additional validation is useful during the import script
writing process but it does add additional processing which
we don't need if we're trying to run lots of import scripts
(such as when we are building a server image).

As an example, the check added in the previous commit adds
1 extra spatial DB query per polling station record imported.
If we're importing data for hundreds of areas, that adds
a non-zero overhead to the import process.

This commit addresses that by adding a 'nochecks' flag which is
off by default (i.e: if we are running a single import script,
we run the checks unless --nochecks is passed). The flag is
passed by default if we are running multiple scripts in bulk
using import.py.

As we add additional validation the checks can be wrapped in
`if self.validation_checks` blocks.