Extend generic BaseCommand
This means system checks for the full host project are run before
addressbase commands. In principle, this means we might perform a
check which is not strictly relevant to the specific command but
in practice it is not a big issue and this makes it easier to pull
the AddressBase code out into a generic module that can be shared
with other projects.