make API and front-end behave consistently if we can't geocode
This handles an edge case I expect we will see a lot less
now that we never fall back to mapit, but it is useful to
bring the two into line. This allows us to populate the
`council` and `polling_station` keys in the json response
but return a null `postcode_location` if we know the
polling station from ResidentialAddress data but can't
geocode the source point for some reason.