present 3x example responses as if they were 3 endpoints
This is less correct than presenting 3 example
responses for one endpoint but it allows us to
present more context info around the response.

I _think_ this makes it clearer on an intial read
(without diving in to the schema) what the major
classes of response are and why they exist.