Allow a shp_encoding parameter to be passed in imports
This commit creates a ShpMixin similar to the CsvMixin in the
base_importers module. To avoid key collisions in the options dict the
keys have been renamed in the CsvMixin as well.

This is needed because the The PyShp Library uses a default encoding of
utf-8, which is (probably) not the defacto encoding for UK local authorities.
This is because ESRI products use a latin-1 encoding by default and are
where a lot of our files will come from.