Import Script for Highland
The UPRNs added to 'bad_uprns' in misc_fixes.py have not been all been checked
as bad, rather it is all uprns that are >16km from the polling place.
This reduces the number of address pickers by throwing out some actual bad
uprns, but it probably includes some good uprns too. As such it would be good to
revert this after May 23rd.