Set SRID after construction if building Geometry from GeoJSON
As of Django 2, constructing a GEOSGeometry() from GeoJSON
automatically assumes SRID=4326 unless the GeoJSON explicitly
includes a CRS member. See:
https://docs.djangoproject.com/en/2.2/releases/2.0/#django-contrib-gis

This is more technically correct, but we do sometimes need to
deal with GeoJSON which uses 27700 and doesn't have a CRS set
(for various reasons), so we always want to set the SRID.

However, setting it in the constructor now throws
ValueError: Input geometry already has SRID: 4326

This commit changes it so we either construct from WKT
(if possible) so we can set the SRID with the constructor or,
if we have to construct from GeoJSON, set the SRID explicitly
after constructing the GEOSGeometry.