Dashboard (#2111)
Add (the start of) a dashboard app

It's intended to be used to quickly check some aspects of loaded data, but not
to be a one-stop shop for all data checking.

It's got an is_staff restriction on it, and currently provides:

* a list of councils
* a page per council
* a page per postcode
* a page per polling station

On the council pages:

* the DataQuality.report
* top ten furthest distances, with links to address_view and the distance in
  metres
* largest diameter postcodes, with diameter in metres and a link to a new
  postcode overview view
* list of polling stations, and whether each has a location

The postcode overview view shows all the properties in that postcode on a map,
with their respective polling stations, coloured by station. useful for
sanity-checking large postcode areas before digging in further with postgis.