Add a flag to use postcode centroids to georeference stations
We generally don't want this behavior in production because it introduced more
problems than it solved - "The church is at the other end of the street!" etc.
However when checking that polling districts and their stations look plausible
it is really useful. So add the flag:
	$ manage.py import_place -p
and the postcode centroids will be used to create station points.