A command to make a csv of UPRNs and councils
The purpose of this is to create a file which can be used to build a
table of the same. At the moment we use addressbase_onsud for this.
We want to get rid of addressbase_onsud because it introduces one more
source of uprns and addresses, this:
  - makes reasoning about issues when writing import scripts harder
  - means we don't ignore UPRNs in addressbase that are missing from
    ONSUD.

We're doing this dance with outputting a file to be imported and then
importing it into a different table to avoid having to rewrite to much,
as that's what we already do but with ONSUD. It's worth noting the
the reason we don't update the addressbase_address table with UPRNs is
it's really slow!