Maintain public API for council contact details
This is a temp solution that means we can punt changing
the public API (including devs.DC) until we're ready to make all
the changes at once.

For the moment, just map the old field names on to the council
contact details.

This leaves a problem of not being able to use the registration details
at the moment, but this isn't making the public interface any worse
than it currently is.