Adapt AddressList to match stations to uprns
This alters AddressList from data_collection/data_types.py for the
data_importers app. Instead of doing a lot of address cleaning, its new
function is to assign a polling station id to each uprn in the
UprnToCouncil model.

Records are discarded if:
  - They don't have a uprn (self.append won't add them)
  - The uprn is not assigned to the council gss code in UprnToCouncil
  - There is more than one record with the same uprn and they are
    assigned to different polling station ids.
  - The postcode on the record differs to the postcode in addressbase.
Currently there is no fuzzymatching on the addresses between the record and
addressbase.

The AssignPollingStationMixin is abstracted out to be used with the
DistrictSet and StationSet classes.