Update routing based on new data_importers app
At a high level these changes remove the 'multiple councils' scenario.
This is because we don't have to ignore postcodes that are spread over
more than one local authority area. Instead in this situation we will
fall back to an address picker.

The ResidentialAddress table will also be removed, and the information
it contained is now accessed using methods on the AddressManager class.