Add a data_importers app with DataQuality Model
This will be the new app for importing information about where to vote
based on information from councils. The new approach will be to assign
polling station ids to a list of UPRNs derived from addressbase.
There will be a lot of duplication of the data_collection app, and
commits should highlight changes.

As well as initialising the data_importers app this commit copies over
the DataQuality model from data_collection but removes most of the
fields. The existing fields are needed to generate the data quality
report and dashboard views.