Add 'saved' flag to CustomSet in data_types
Now if you call get_polling_station_lookup without having saved records
to the db a 'RecordNotSavedException' will be raised.

This is needed becasue get_polling_station_lookup uses st_contains in
the db. The spatial lookup is done in the db becasue building the index
is all taken care of, and it's probably quicker than doing it using
django's spatial objects.