Fixing geocoders
AddressBaseGeocoderAdaptor should just return a AddressBaseGeocoder and
not try to do anything cleverer.

This stops us from catching addressbase records that aren't in the
UprnToCouncil table, these are very few, but we should either check
somewhere else, or delete addressbase records that don't make it past
the point in council polygon lookup.