Add a PostcodeBuilder class for data_finder tests
The aim of this class is to make it easier to define, maintain and
reason about the multiple ways a postcode can exist in the db that we
want to test for. The aim is that, combined with various factory classes
that it depends on, it should replace a lot of the json fixtures which
can become unwieldly to read and alter. This should also make it easier
to understand what the tests are meant to be testing.