Add CouncilGeography model.
This deletes the area field from council and adds a new model to store
the shape associated with a council. This is to seperate out the idea of
a council that can persist through several boundary changes. This is in
line with how we model things in EveryElection, so a similar abstraction
seems reasonable here.

At the moment this is being added as a 'OneToOne' field so we don't have
to deal with the possibility of storing multiple geographies per
council. EE is much better placed to handle that.