Change Dashboard query
Calculate multipoint length only accountng for uprns with a polling
station. This is because we only care if a uprn has the wrong postcode
or lat/lon _if_ we have a polling station for it. Otherwise we'll just
be saying 'don't know - ask you council' - user won't get a map in these
cases either.

In the maxdistance query we can drop the onspd table because we only
have to consider addresses which have a lat/lon from addressbase.