Warn on stations with same location, different postcode
It's not uncommon to find two polling stations with the same lat/lon in
the council file, but different addresses - i.e. one has the wrong
lat/lon. However it's also common to have multiple stations at the same
location, e.g. different rooms in a school. In order to catch the
stations that have been given the wrong location in error, this check
compares the distance between the station to be added to the stationset
and all the stations added so far.

This check won't happen when the --nochecks flag is passed.

This commit also refactors the check_station_point method now there are
two checks, both of which require a location.

The qgis filter expression line allows the user to quickly identify the
stations in question by applying a filter to the polling stations layer.