Improve routing helper performance
We want to know all the councils for a postcode, so that we know if the
postcode is split over council boundaries. Previously we were doing a
lookup per UPRN in the postcode. In postcodes with lots of UPRNs this is
very slow. This means there is one query for the whole postcode.