Query db once per council per postcode-API call
If not all addresses in a postcode have the same station we show an
array of addresses with details including council and station.
Previously we would query the db once per address to get the council,
(using the council property on the address model). This is a fairly
hacky way round this, where we set the property on the address objects
that we've got via the RoutingHelper object when we check whether the
postcode is split by councils.

This is probably not the right way to do it, but works-for-now™.