Pull string construction into template for the directions time/distance
This replaces collections.namedtuple with typed.NamedTuple, which has
the same functionality, but is easier to document with types and
comments.

Time and distance are now passed around in natural units (seconds and
metres) with utility properties on Directions to convert to minutes and
miles.

We can then use `…['value']` instead of `…['text']` when pulling these
values out of the Google directions response.

Finally, the {% blocktrans %} for summarising the directions is cased on
the mode, with two strings to translate; one for walking and the other
for driving.