Fix feedback form JS
This does a few things:
* We submit form data on any YES/NO button click, in case the user
doesn't bother with clicking the 'submit' button. This is now handled by
the 'post_partial_form' function.
* Now there are two fields that might be updated, we can't use
create_or_update, because we could overwrite the other value. Instead we
need to check whether a feedback model exists already or not, and only
update the new value.
* If we're 'reporting a problem' hide the 'more likely to vote' form.
* Only show the 'tell us more' box when both the other options have been
entered.

Co-authored-by: Sym Roe <sym.roe@talusdesign.co.uk>