Only show NI CustomFinder when there is no NI data
We're now recieving data from EONI. They want results from our site to
mirror those from their site. So we shouldn't show a custom finder if we
have EONI data in our db. Instead our fallback (say someone enters a
postcode that is in ONSPD but not the EONI data export) should be to
direct users to contact EONI - not use their finder.
However it is possible that we won't have EONI data for some reason. If
that is the case we probably do want to direct users to use the custom
finder. This reasoning applies to API consumers too.