ImportScript can now adapt an existing script
This uses the `ast` module and some line-based wrangling to replace
particular values in an existing script.

We can't easily do this with line-based pattern matching as sometimes our
addresses_name/stations_name assignment statements are split over
multiple lines due to black formatting, and processing would be liable
to be fiddly and/or fragile (probably both).