Change db routing to read/write
Given we're now going to use RDS as a principal database, and then
logical replication to the on instance databases, we can change out
database routing to all reads from local, and all writes to prinicipal.

This feels a little hairy, because there'll be some lag between the
write being sent to RDS, replicated to local, and then appearing for the
user. However pushing this should let us test it a bit in development.