Use named Indexes in Addressbase
This is so that we can be specific about the indexes to drop and
recreate when setting up database replication.

We also create a foreign key constraint after completing replication.
This is dropped thanks to the CASCADE when dropping the PKs.
Naming of the FK is handled by django in db/backends/base/schema.py
The unique hash is generated from the table name and column names. The
FK shouldn't be explicitly referenced by any django generated SQL, so we
can recreate it with any (unique) name. Hopefully.