Fix CSV fixtures
Something had removed the trailing whitespace from these, and that in
turn broke the tests. This might be a python BC between versions,
unsure.

Either way, be really careful about touching these files, as most
editors will strip trailing whitespace (the pre-commit in this repo does
the same). The whitespace is required