Add --include-past-elections flag to import scripts
This goal of this commit is to make sure import scripts only run when
the 'elections' attribute on an Importer class contains a date that is
today or in the future.
Adding the new --include-past-elections flag overrides the checks for
the elections attribute, and that at least some dates are in the future.

There's a limitation in that the new covers_current_elections method is
assuming the contents of the elections attributes are isodate-like,
rather than DC Election ID like. This reflects how we use them, rather
than how the use was perhaps envisioned.

The --include-past-elections has also been exposed in the import.py
command, which is for running multiple import scripts, and therefore a
likely place where the functionality might be needed.