Use RDS for upload and file models
This is to deal with 'read your writes' consistency issues resulting in
replication lag. This is unlikely to be acutally seen by users, but it's
possible for our s3 trigger posting to the upload endpoint to trigger
error emails due to a missing file_set on the local db, which exists on
the RDS instance.