Match strings to strings, not datetimes
We should compare string-y dates here not datetime dates.
This commit also makes sure there is a 'NEXT_CHARISMATIC_ELECTION_DATE'
variable available in views, because it is needed by the advance voting
logic in the polling_stations_known.html fragment. Obviously this only
works if the AVS is set for the next charismatic election, but that's a
moot point in may 2023...