Assume address location on api address view
When we get the EE wrapper we were only using the addresss location when
there wasn't a single station for the postcode. However this meant when
all addresses vote at a single station, but they are split across wards,
we weren't returning the correct ward for an address that was in a
different ward to the postcode centroid. With this change we'll still
omit an address picker when there should be one, but this endpoint will
be correct.