Change footer headings to p tags
The HTML footer element isn't a section element, meaning that headers in
a footer are considered part of the heading structure of the containing
section.

That is, before this change we would have a content structure like:

- Find your polling station
  - Where Do I Vote
  - Made by Democracy Club

These last two aren't useful headers, but we want them to look large on
the page.

This change uses `ds-h4` to make them look the same, but removes them
from the content structure