Replace model authentication with hard coded tokens
This commit removes the model based authendiation with a hard coded list
of tokens.

There are two types of token: a normal authenticated user token and a
super user token. The first is read only (in the current implementation)
and the second can write content. At the moment, the only write possible
is file uploads.

This change disables all non-authenticated API views, meaning that there
will be no public API for this project any more. The devs.DC API should
be used to present endpoints to the world.