Put URL on newline, don't add full stop
Some email clients include the full stop in the link. As the end of the URL is the token, this results in URLs like:

`?token=123456.` `123456.` isn't a valid token (`123456` is), so authentication fails.

This change puts the URL on a new line and remove the full stop. This should prevent an (imagined) case where email clients break the link on a new line.