Add eoni specific context to BasePollingStationView
ni_out_of_cycle_station should be true when we know the station, but
when there is no election. When there's an election we just want to
handle things normally.

ni_elected_rep_type is just helper for showing the text that is specific
to which scheme we have live.

SHOW_EONI_STATIONS_ALL_THE_TIME is a flag in settings to control whether
we show the stations out of cycle or not.