WIP intermediate refactoring step
The objective here is I want to split EveryElectionWrapper
into two classes:
- EEFetcher (which requests data from EE)
- EEWrapper (which provides some higher level convenience
  functions over the data)
To start with, I'm putting this new code in a file called
ee2.py (great name) and leaving the existing classes in
every_election.py as wrappers over the new classes.
This makes it easier to start on the new classes without
having to update every call site just yet.
We will clean this up later.