Creates a file that looks like addressbase_cleaned.csv from onspd
This gets all the postcodes that aren't in addressbase, aren't large
users, and, are on mainland GB (i.e. no Northern Ireland, Channel
Islands, Isle of Man) and creates a csv that has the same shape as
addressbase_cleaned.csv. However there's a dummy uprn in the form:
  ONSPD:TE11ST
Also the address field is blank. This should let us chuck it into the
data-baker pipeline to end up in the parquet.